requirements.txt
pytesseract
Pillow
pdf2image
pyzbar
typing-extensions
pytest

